refact: rename capture directory to recovery
