+++ tests/MoneticoTest.php
@@ -1,7 +1,7 @@
 <?php
 
 use Carbon\Carbon;
-use DansMaCulotte\Monetico\Capture\Recovery;
+use DansMaCulotte\Monetico\Recovery\Recovery;
 use DansMaCulotte\Monetico\Monetico;
 use \DansMaCulotte\Monetico\Exceptions\Exception;
 use DansMaCulotte\Monetico\Payment\Payment;
